Move PlatformSpecificLinkedLibs and implementations to common and rename
