+++ BizHawk.Common/BizHawk.Common.csproj
@@ -84,6 +84,7 @@
     <Compile Include="Log.cs" />
     <Compile Include="MruStack.cs" />
     <Compile Include="NDBDatabase.cs" />
+    <Compile Include="PlatformLinkedLibSingleton.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="QuickCollections.cs" />
     <Compile Include="Serializer.cs" />
